Initial checkin of Robot Camera
This programs a ESP32-CAM board to act as a webcam in conjunction with the BrainSimulator RobotCamera module. You can use a bluetooth connection to connect the board to wifi. Then the board creates a webserver and every time the page (at the ip address) is accessed, it returns a jpeg of the current camera content. It is currently fixed at 320x240 resoluation.
There is a .sln file which works with Visual Micro but it seems to work equally well in the Arduino IDE.